+++ src/main/java/org/opensearch/ad/client/AnomalyDetectionClient.java
@@ -12,6 +12,10 @@ import org.opensearch.ad.transport.GetAnomalyDetectorResponse;
 import org.opensearch.common.action.ActionFuture;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.timeseries.transport.GetConfigRequest;
+import org.opensearch.timeseries.transport.SuggestConfigParamRequest;
+import org.opensearch.timeseries.transport.SuggestConfigParamResponse;
+import org.opensearch.timeseries.transport.ValidateConfigRequest;
+import org.opensearch.timeseries.transport.ValidateConfigResponse;
 
 /**
  * A client to provide interfaces for anomaly detection functionality. This will be used by other plugins.
@@ -71,4 +75,39 @@ public interface AnomalyDetectionClient {
      */
     void getDetectorProfile(GetConfigRequest profileRequest, ActionListener<GetAnomalyDetectorResponse> listener);
 
+    /**
+     * Validate anomaly detector - refer to https://opensearch.org/docs/latest/observing-your-data/ad/api/#validate-detector
+     * @param validateRequest request to validate the detector configuration
+     * @return ActionFuture of ValidateConfigResponse
+     */
+    default ActionFuture<ValidateConfigResponse> validateAnomalyDetector(ValidateConfigRequest validateRequest) {
+        PlainActionFuture<ValidateConfigResponse> actionFuture = PlainActionFuture.newFuture();
+        validateAnomalyDetector(validateRequest, actionFuture);
+        return actionFuture;
+    }
+
+    /**
+     * Validate anomaly detector - refer to https://opensearch.org/docs/latest/observing-your-data/ad/api/#validate-detector
+     * @param validateRequest request to validate the detector configuration
+     * @param listener a listener to be notified of the result
+     */
+    void validateAnomalyDetector(ValidateConfigRequest validateRequest, ActionListener<ValidateConfigResponse> listener);
+
+    /**
+     * Suggest anomaly detector parameters - refer to https://opensearch.org/docs/latest/observing-your-data/ad/api/#suggest-detector
+     * @param suggestRequest request to suggest detector configuration parameters
+     * @return ActionFuture of SuggestConfigParamResponse
+     */
+    default ActionFuture<SuggestConfigParamResponse> suggestAnomalyDetector(SuggestConfigParamRequest suggestRequest) {
+        PlainActionFuture<SuggestConfigParamResponse> actionFuture = PlainActionFuture.newFuture();
+        suggestAnomalyDetector(suggestRequest, actionFuture);
+        return actionFuture;
+    }
+
+    /**
+     * Suggest anomaly detector parameters - refer to https://opensearch.org/docs/latest/observing-your-data/ad/api/#suggest-detector
+     * @param suggestRequest request to suggest detector configuration parameters
+     * @param listener a listener to be notified of the result
+     */
+    void suggestAnomalyDetector(SuggestConfigParamRequest suggestRequest, ActionListener<SuggestConfigParamResponse> listener);
 }