+++ src/main/java/org/opensearch/ad/client/AnomalyDetectionNodeClient.java
@@ -13,10 +13,16 @@ import org.opensearch.ad.transport.GetAnomalyDetectorAction;
 import org.opensearch.ad.transport.GetAnomalyDetectorResponse;
 import org.opensearch.ad.transport.SearchAnomalyDetectorAction;
 import org.opensearch.ad.transport.SearchAnomalyResultAction;
+import org.opensearch.ad.transport.SuggestAnomalyDetectorParamAction;
+import org.opensearch.ad.transport.ValidateAnomalyDetectorAction;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.action.ActionResponse;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.timeseries.transport.GetConfigRequest;
+import org.opensearch.timeseries.transport.SuggestConfigParamRequest;
+import org.opensearch.timeseries.transport.SuggestConfigParamResponse;
+import org.opensearch.timeseries.transport.ValidateConfigRequest;
+import org.opensearch.timeseries.transport.ValidateConfigResponse;
 import org.opensearch.transport.client.Client;
 
 public class AnomalyDetectionNodeClient implements AnomalyDetectionClient {
@@ -47,6 +53,16 @@ public class AnomalyDetectionNodeClient implements AnomalyDetectionClient {
         this.client.execute(GetAnomalyDetectorAction.INSTANCE, profileRequest, getAnomalyDetectorResponseActionListener(listener));
     }
 
+    @Override
+    public void validateAnomalyDetector(ValidateConfigRequest validateRequest, ActionListener<ValidateConfigResponse> listener) {
+        this.client.execute(ValidateAnomalyDetectorAction.INSTANCE, validateRequest, validateConfigResponseActionListener(listener));
+    }
+
+    @Override
+    public void suggestAnomalyDetector(SuggestConfigParamRequest suggestRequest, ActionListener<SuggestConfigParamResponse> listener) {
+        this.client.execute(SuggestAnomalyDetectorParamAction.INSTANCE, suggestRequest, suggestConfigResponseActionListener(listener));
+    }
+
     // We need to wrap AD-specific response type listeners around an internal listener, and re-generate the response from a generic
     // ActionResponse. This is needed to prevent classloader issues and ClassCastExceptions when executed by other plugins.
     // Additionally, we need to inject the configured NamedWriteableRegistry so NamedWriteables (present in sub-fields of
@@ -60,6 +76,31 @@ public class AnomalyDetectionNodeClient implements AnomalyDetectionClient {
         }, listener::onFailure);
         ActionListener<GetAnomalyDetectorResponse> actionListener = wrapActionListener(internalListener, actionResponse -> {
             GetAnomalyDetectorResponse response = GetAnomalyDetectorResponse
+                .fromActionResponse(actionResponse, this.namedWriteableRegistry);
+            return response;
+        });
+        return actionListener;
+    }
+
+    private ActionListener<ValidateConfigResponse> validateConfigResponseActionListener(ActionListener<ValidateConfigResponse> listener) {
+        ActionListener<ValidateConfigResponse> internalListener = ActionListener.wrap(validateConfigResponse -> {
+            listener.onResponse(validateConfigResponse);
+        }, listener::onFailure);
+        ActionListener<ValidateConfigResponse> actionListener = wrapActionListener(internalListener, actionResponse -> {
+            ValidateConfigResponse response = ValidateConfigResponse.fromActionResponse(actionResponse, this.namedWriteableRegistry);
+            return response;
+        });
+        return actionListener;
+    }
+
+    private ActionListener<SuggestConfigParamResponse> suggestConfigResponseActionListener(
+        ActionListener<SuggestConfigParamResponse> listener
+    ) {
+        ActionListener<SuggestConfigParamResponse> internalListener = ActionListener.wrap(suggestConfigResponse -> {
+            listener.onResponse(suggestConfigResponse);
+        }, listener::onFailure);
+        ActionListener<SuggestConfigParamResponse> actionListener = wrapActionListener(internalListener, actionResponse -> {
+            SuggestConfigParamResponse response = SuggestConfigParamResponse
                 .fromActionResponse(actionResponse, this.namedWriteableRegistry);
             return response;
         });